Create a soname link to the shared library if it doesn't exist
Shared libraries can have the DT_SONAME property set which refers to
the logical name of the library and is the name looked for by the
dynamic loader.  Usually this is the same as the filename but when
it is different use objdump -p to grab the SONAME value and set a link
to create it.