Refactor NarIntegrationTest
It has been some time since I made all these changes so I do not recall
entirely the reason for the change but I believe it was motiviated to
be able to utilize some newer test features provided by an updated
surefire plugin.  However, the original copied a lot of that
implementation into this mojo.  I removed that implementation, added a
dependency to the plugin, and called the plugin from the
NarIntegrationTestMojo class.  One benefit for users of TestNG is that
it acutally displays the name of the test class that is currently
being tested.