Restore deleted code back to the NarManager class
This is a partial undo of commit 8906f94f29d0194deae7ef68d000d43214226171.
The code was deleted because it wasn't used at the time.  Some of the
code to be added will require this code.  It is unclear whether the
code was simply unused or if it became unused because an alternative
approach was used.