Correctly handle wildcards in property values
Some shared library extensions are configured with the value of
"so*".  The wildcard at the end is to presumably catch versioned
so files such as libfile.so.2.3.  For the nar-properties goal
substituting a property with a value containing a wildcard is not
desireable.  Assume the request is for the link name of the
library (which would just be be libfile.so) and drop the wildcard
from the property value when substituting properties.