Create hardlinks rather than softlinks to the so file
When a dependent NAR is a shared library that library can
have multiple names; a link name (xxxx.so) used during linking
and a versioned name or soname (xxxx.so.2) that is used during
execution.  The dependency may not include both filenames.  When
that is the case, the NAR plugin will create those files via a
softlink to the provided library file.

However, after building the project, it is common to want to
package all the native libraries into a jarfile through some
sort of assembly.  If the versioned or soname library needed
for execution is a softlink, the maven assembly plugin (or other
packing plugin) will just copy the link file and not the file
the link is pointing to.  If the file is a hardlink instead of
a softlink a valid library file will be copied into the assembly.