Fix nar-vcprops goal when using snapshot dependencies
With a snapshot dependency the NarDependency.getVersion() returns
a timestamped version but the nar plugin extracts the include and
library files using the NarDependency's base version causing the
goal to fail because the timestamped include and library
directories don't exist.  Use the base version to construct the
include path rather than the (possibly) timestamped version.