Don't override the AOL/default binding for shared/static types
If a project built multiple output types
(executable/none/jni/shared/static) the default and AOL binding
were overridden to shared if there was a shared output type or
to static if there was a static output type.

Change the behavior slightly to not set the shared and/or static
binding type if the nar.properties file already specified a
particular binding.  If no binding is specified the selected
default binding type functions as before.