Add support for Visual Studio 2019
The original method for finding Visual Studio was to look at a
particular registry setting.  That value is no longer being set
with 2019.  Instead use the vswhere executable which has been
packaged with Visual Studio since at least 2017.  If that fails,
continue with the older method of checking the register, then the
environment variables.