Improve CMake detection
Add the ability to configure the name of the CMake executable.
On Windows if the CMake location isn't found in the registry see
if it can be run by relying on the PATH variable.  On Linux search
the cmake executable by prefix.  Some linux distros name the most
recent version of cmake to cmake3.